+++ CleverFerret/src/main/java/com/universalmedialibrary/debug/ui/DebugMenuScreen.kt
@@ -228,50 +228,6 @@ fun DebugMenuScreen(
     }
 }
 
-@Composable
-private fun PerformanceSummaryCard(
-    metrics: PerformanceMetrics,
-    crashCount: Int,
-    errorCount: Int
-) {
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(16.dp),
-        colors = CardDefaults.cardColors(containerColor = MediaColors.BackgroundSurface),
-        shape = RoundedCornerShape(12.dp)
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(16.dp),
-            horizontalArrangement = Arrangement.SpaceEvenly
-        ) {
-            MetricItem(
-                icon = Icons.Default.Memory,
-                label = "Memory",
-                value = "${metrics.memoryUsedMB}/${metrics.memoryMaxMB} MB",
-                color = when {
-                    metrics.memoryPercentUsed > 80 -> MediaColors.Error
-                    metrics.memoryPercentUsed > 60 -> MediaColors.Warning
-                    else -> MediaColors.Success
-                }
-            )
-            MetricItem(
-                icon = Icons.Default.Warning,
-                label = "Crashes",
-                value = "$crashCount",
-                color = if (crashCount > 0) MediaColors.Error else MediaColors.Success
-            )
-            MetricItem(
-                icon = Icons.Default.Error,
-                label = "Errors",
-                value = "$errorCount",
-                color = if (errorCount > 0) MediaColors.Warning else MediaColors.Success
-            )
-        }
-    }
-}
 
 @Composable
 private fun MetricItem(
@@ -307,330 +263,11 @@ private fun MetricItem(
 // TABS
 // =============================================================================
 
-@Composable
-private fun CrashReportsTab(
-    crashes: List<CrashReport>,
-    onCrashClick: (CrashReport) -> Unit,
-    onClearAll: () -> Unit
-) {
-    Column(modifier = Modifier.fillMaxSize()) {
-        if (crashes.isNotEmpty()) {
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp),
-                horizontalArrangement = Arrangement.End
-            ) {
-                TextButton(onClick = onClearAll) {
-                    Icon(Icons.Default.Delete, contentDescription = null)
-                    Spacer(Modifier.width(4.dp))
-                    Text("Clear All")
-                }
-            }
-        }
-        
-        if (crashes.isEmpty()) {
-            EmptyState(
-                icon = Icons.Default.CheckCircle,
-                message = "No crashes recorded",
-                color = MediaColors.Success
-            )
-        } else {
-            LazyColumn(
-                modifier = Modifier.fillMaxSize(),
-                contentPadding = PaddingValues(16.dp),
-                verticalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                items(crashes, key = { it.id }) { crash ->
-                    CrashReportCard(
-                        crash = crash,
-                        onClick = { onCrashClick(crash) }
-                    )
-                }
-            }
-        }
-    }
-}
 
-@Composable
-private fun CrashReportCard(
-    crash: CrashReport,
-    onClick: () -> Unit
-) {
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable(onClick = onClick),
-        colors = CardDefaults.cardColors(
-            containerColor = MediaColors.Error.copy(alpha = 0.1f)
-        ),
-        shape = RoundedCornerShape(8.dp)
-    ) {
-        Row(
-            modifier = Modifier.padding(12.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Icon(
-                Icons.Default.Error,
-                contentDescription = null,
-                tint = MediaColors.Error,
-                modifier = Modifier.size(32.dp)
-            )
-            Spacer(Modifier.width(12.dp))
-            Column(modifier = Modifier.weight(1f)) {
-                Text(
-                    text = crash.exceptionClass.substringAfterLast('.'),
-                    fontWeight = FontWeight.Bold,
-                    color = MediaColors.TextPrimary
-                )
-                Text(
-                    text = crash.message,
-                    fontSize = 12.sp,
-                    color = MediaColors.TextSecondary,
-                    maxLines = 2,
-                    overflow = TextOverflow.Ellipsis
-                )
-                Text(
-                    text = formatTimestamp(crash.timestamp),
-                    fontSize = 10.sp,
-                    color = MediaColors.TextSecondary
-                )
-            }
-            Icon(
-                Icons.Default.ChevronRight,
-                contentDescription = "View details",
-                tint = MediaColors.TextSecondary
-            )
-        }
-    }
-}
 
-@Composable
-private fun ErrorLogsTab(
-    errors: List<ErrorLog>,
-    onErrorClick: (ErrorLog) -> Unit,
-    onClearAll: () -> Unit
-) {
-    Column(modifier = Modifier.fillMaxSize()) {
-        if (errors.isNotEmpty()) {
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp),
-                horizontalArrangement = Arrangement.End
-            ) {
-                TextButton(onClick = onClearAll) {
-                    Icon(Icons.Default.Delete, contentDescription = null)
-                    Spacer(Modifier.width(4.dp))
-                    Text("Clear All")
-                }
-            }
-        }
-        
-        if (errors.isEmpty()) {
-            EmptyState(
-                icon = Icons.Default.CheckCircle,
-                message = "No errors logged",
-                color = MediaColors.Success
-            )
-        } else {
-            LazyColumn(
-                modifier = Modifier.fillMaxSize(),
-                contentPadding = PaddingValues(16.dp),
-                verticalArrangement = Arrangement.spacedBy(4.dp)
-            ) {
-                items(errors.reversed(), key = { it.id }) { error ->
-                    ErrorLogItem(
-                        error = error,
-                        onClick = { onErrorClick(error) }
-                    )
-                }
-            }
-        }
-    }
-}
 
-@Composable
-private fun ErrorLogItem(
-    error: ErrorLog,
-    onClick: () -> Unit
-) {
-    val color = when (error.level) {
-        LogLevel.ERROR -> MediaColors.Error
-        LogLevel.WARNING -> MediaColors.Warning
-        LogLevel.INFO -> MediaColors.AccentPrimary
-    }
-    
-    Surface(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable(onClick = onClick),
-        color = color.copy(alpha = 0.1f),
-        shape = RoundedCornerShape(4.dp)
-    ) {
-        Row(
-            modifier = Modifier.padding(8.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Box(
-                modifier = Modifier
-                    .size(8.dp)
-                    .clip(CircleShape)
-                    .background(color)
-            )
-            Spacer(Modifier.width(8.dp))
-            Column(modifier = Modifier.weight(1f)) {
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    Text(
-                        text = error.tag,
-                        fontSize = 10.sp,
-                        fontWeight = FontWeight.Bold,
-                        fontFamily = FontFamily.Monospace,
-                        color = color
-                    )
-                    Text(
-                        text = formatTimestamp(error.timestamp),
-                        fontSize = 10.sp,
-                        color = MediaColors.TextSecondary
-                    )
-                }
-                Text(
-                    text = error.message,
-                    fontSize = 12.sp,
-                    color = MediaColors.TextPrimary,
-                    maxLines = 2,
-                    overflow = TextOverflow.Ellipsis
-                )
-            }
-        }
-    }
-}
 
-@Composable
-private fun PerformanceTab(metrics: PerformanceMetrics) {
-    LazyColumn(
-        modifier = Modifier.fillMaxSize(),
-        contentPadding = PaddingValues(16.dp),
-        verticalArrangement = Arrangement.spacedBy(16.dp)
-    ) {
-        item {
-            Card(
-                colors = CardDefaults.cardColors(containerColor = MediaColors.BackgroundSurface),
-                shape = RoundedCornerShape(12.dp)
-            ) {
-                Column(modifier = Modifier.padding(16.dp)) {
-                    Text(
-                        "Memory Usage",
-                        fontWeight = FontWeight.Bold,
-                        color = MediaColors.TextPrimary
-                    )
-                    Spacer(Modifier.height(12.dp))
-                    
-                    LinearProgressIndicator(
-                        progress = { metrics.memoryPercentUsed / 100f },
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .height(8.dp)
-                            .clip(RoundedCornerShape(4.dp)),
-                        color = when {
-                            metrics.memoryPercentUsed > 80 -> MediaColors.Error
-                            metrics.memoryPercentUsed > 60 -> MediaColors.Warning
-                            else -> MediaColors.Success
-                        },
-                        trackColor = MediaColors.BackgroundElevated
-                    )
-                    
-                    Spacer(Modifier.height(8.dp))
-                    Row(
-                        modifier = Modifier.fillMaxWidth(),
-                        horizontalArrangement = Arrangement.SpaceBetween
-                    ) {
-                        Text(
-                            "${metrics.memoryUsedMB} MB used",
-                            fontSize = 12.sp,
-                            color = MediaColors.TextSecondary
-                        )
-                        Text(
-                            "${metrics.memoryMaxMB} MB max",
-                            fontSize = 12.sp,
-                            color = MediaColors.TextSecondary
-                        )
-                    }
-                }
-            }
-        }
-        
-        item {
-            Text(
-                "Last updated: ${formatTimestamp(metrics.lastUpdated)}",
-                fontSize = 12.sp,
-                color = MediaColors.TextSecondary
-            )
-        }
-    }
-}
 
-@Composable
-private fun FlagsTab(
-    flags: Map<String, Boolean>,
-    onToggleFlag: (String, Boolean) -> Unit
-) {
-    LazyColumn(
-        modifier = Modifier.fillMaxSize(),
-        contentPadding = PaddingValues(16.dp),
-        verticalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        item {
-            Text(
-                "Feature Flags",
-                fontWeight = FontWeight.Bold,
-                color = MediaColors.TextPrimary,
-                modifier = Modifier.padding(bottom = 8.dp)
-            )
-        }
-        
-        items(flags.entries.toList()) { (key, value) ->
-            Card(
-                colors = CardDefaults.cardColors(containerColor = MediaColors.BackgroundSurface),
-                shape = RoundedCornerShape(8.dp)
-            ) {
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(16.dp),
-                    horizontalArrangement = Arrangement.SpaceBetween,
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Column(modifier = Modifier.weight(1f)) {
-                        Text(
-                            text = key.replace("_", " ").lowercase()
-                                .replaceFirstChar { it.uppercase() },
-                            fontWeight = FontWeight.Medium,
-                            color = MediaColors.TextPrimary
-                        )
-                        Text(
-                            text = key,
-                            fontSize = 10.sp,
-                            fontFamily = FontFamily.Monospace,
-                            color = MediaColors.TextSecondary
-                        )
-                    }
-                    Switch(
-                        checked = value,
-                        onCheckedChange = { onToggleFlag(key, it) },
-                        colors = SwitchDefaults.colors(
-                            checkedThumbColor = MediaColors.AccentPrimary,
-                            checkedTrackColor = MediaColors.AccentPrimary.copy(alpha = 0.5f)
-                        )
-                    )
-                }
-            }
-        }
-    }
-}
 
 @Composable
 private fun NotificationsTab(
@@ -675,99 +312,7 @@ private fun NotificationsTab(
     }
 }
 
-@Composable
-private fun ToolsTab(
-    onExportLogs: () -> Unit,
-    onSimulateCrash: () -> Unit
-) {
-    LazyColumn(
-        modifier = Modifier.fillMaxSize(),
-        contentPadding = PaddingValues(16.dp),
-        verticalArrangement = Arrangement.spacedBy(12.dp)
-    ) {
-        item {
-            Text(
-                "Developer Tools",
-                fontWeight = FontWeight.Bold,
-                color = MediaColors.TextPrimary,
-                modifier = Modifier.padding(bottom = 8.dp)
-            )
-        }
-        
-        item {
-            Card(
-                colors = CardDefaults.cardColors(containerColor = MediaColors.BackgroundSurface),
-                shape = RoundedCornerShape(12.dp)
-            ) {
-                Column(modifier = Modifier.padding(16.dp)) {
-                    ToolButton(
-                        icon = Icons.Default.Download,
-                        title = "Export All Logs",
-                        description = "Export crash reports, errors, and metrics to JSON",
-                        onClick = onExportLogs
-                    )
-                    
-                    HorizontalDivider(
-                        modifier = Modifier.padding(vertical = 12.dp),
-                        color = MediaColors.BackgroundElevated
-                    )
-                    
-                    ToolButton(
-                        icon = Icons.Default.Warning,
-                        title = "Simulate Crash",
-                        description = "Trigger a test crash to verify crash reporting",
-                        onClick = onSimulateCrash,
-                        destructive = true
-                    )
-                }
-            }
-        }
-    }
-}
 
-@Composable
-private fun ToolButton(
-    icon: ImageVector,
-    title: String,
-    description: String,
-    onClick: () -> Unit,
-    destructive: Boolean = false
-) {
-    val color = if (destructive) MediaColors.Error else MediaColors.TextPrimary
-    
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable(onClick = onClick)
-            .padding(vertical = 8.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Icon(
-            imageVector = icon,
-            contentDescription = null,
-            tint = color,
-            modifier = Modifier.size(24.dp)
-        )
-        Spacer(Modifier.width(16.dp))
-        Column(modifier = Modifier.weight(1f)) {
-            Text(
-                text = title,
-                fontWeight = FontWeight.Medium,
-                color = color
-            )
-            Text(
-                text = description,
-                fontSize = 12.sp,
-                color = MediaColors.TextSecondary
-            )
-        }
-        Icon(
-            Icons.Default.ChevronRight,
-            contentDescription = null,
-            tint = MediaColors.TextSecondary
-        )
-    }
-}
 
 @Composable
 private fun EmptyState(
@@ -802,222 +347,8 @@ private fun EmptyState(
 // DIALOGS
 // =============================================================================
 
-@Composable
-private fun BugReportDialog(
-    onDismiss: () -> Unit,
-    onSubmit: (title: String, description: String, steps: String) -> Unit
-) {
-    var title by remember { mutableStateOf("") }
-    var description by remember { mutableStateOf("") }
-    var steps by remember { mutableStateOf("") }
-    
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = {
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                Icon(
-                    Icons.Default.BugReport,
-                    contentDescription = null,
-                    tint = MediaColors.AccentPrimary
-                )
-                Spacer(Modifier.width(8.dp))
-                Text("Submit Bug Report")
-            }
-        },
-        text = {
-            Column(
-                verticalArrangement = Arrangement.spacedBy(12.dp)
-            ) {
-                OutlinedTextField(
-                    value = title,
-                    onValueChange = { title = it },
-                    label = { Text("Title") },
-                    modifier = Modifier.fillMaxWidth(),
-                    singleLine = true
-                )
-                OutlinedTextField(
-                    value = description,
-                    onValueChange = { description = it },
-                    label = { Text("Description") },
-                    modifier = Modifier.fillMaxWidth(),
-                    minLines = 3
-                )
-                OutlinedTextField(
-                    value = steps,
-                    onValueChange = { steps = it },
-                    label = { Text("Steps to Reproduce") },
-                    modifier = Modifier.fillMaxWidth(),
-                    minLines = 2
-                )
-            }
-        },
-        confirmButton = {
-            Button(
-                onClick = { onSubmit(title, description, steps) },
-                enabled = title.isNotBlank() && description.isNotBlank(),
-                colors = ButtonDefaults.buttonColors(
-                    containerColor = MediaColors.AccentPrimary
-                )
-            ) {
-                Text("Submit")
-            }
-        },
-        dismissButton = {
-            TextButton(onClick = onDismiss) {
-                Text("Cancel")
-            }
-        },
-        containerColor = MediaColors.BackgroundSurface
-    )
-}
 
-@Composable
-private fun CrashDetailsDialog(
-    crash: CrashReport,
-    onDismiss: () -> Unit
-) {
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = {
-            Text(
-                crash.exceptionClass.substringAfterLast('.'),
-                fontWeight = FontWeight.Bold
-            )
-        },
-        text = {
-            Column(
-                modifier = Modifier.verticalScroll(rememberScrollState()),
-                verticalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                Text(
-                    "Message:",
-                    fontWeight = FontWeight.Bold,
-                    color = MediaColors.TextSecondary,
-                    fontSize = 12.sp
-                )
-                Text(crash.message, color = MediaColors.TextPrimary)
-                
-                Text(
-                    "Thread:",
-                    fontWeight = FontWeight.Bold,
-                    color = MediaColors.TextSecondary,
-                    fontSize = 12.sp
-                )
-                Text(crash.threadName, color = MediaColors.TextPrimary)
-                
-                Text(
-                    "Stack Trace:",
-                    fontWeight = FontWeight.Bold,
-                    color = MediaColors.TextSecondary,
-                    fontSize = 12.sp
-                )
-                Surface(
-                    color = MediaColors.Background,
-                    shape = RoundedCornerShape(4.dp)
-                ) {
-                    Text(
-                        text = crash.stackTrace,
-                        fontFamily = FontFamily.Monospace,
-                        fontSize = 10.sp,
-                        color = MediaColors.TextPrimary,
-                        modifier = Modifier.padding(8.dp)
-                    )
-                }
-                
-                Text(
-                    "Device: ${crash.deviceInfo.manufacturer} ${crash.deviceInfo.model}",
-                    fontSize = 10.sp,
-                    color = MediaColors.TextSecondary
-                )
-                Text(
-                    "Android: ${crash.deviceInfo.androidVersion} (API ${crash.deviceInfo.sdkVersion})",
-                    fontSize = 10.sp,
-                    color = MediaColors.TextSecondary
-                )
-            }
-        },
-        confirmButton = {
-            TextButton(onClick = onDismiss) {
-                Text("Close")
-            }
-        },
-        containerColor = MediaColors.BackgroundSurface
-    )
-}
 
-@Composable
-private fun ErrorDetailsDialog(
-    error: ErrorLog,
-    onDismiss: () -> Unit
-) {
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = {
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                val color = when (error.level) {
-                    LogLevel.ERROR -> MediaColors.Error
-                    LogLevel.WARNING -> MediaColors.Warning
-                    LogLevel.INFO -> MediaColors.AccentPrimary
-                }
-                Box(
-                    modifier = Modifier
-                        .size(12.dp)
-                        .clip(CircleShape)
-                        .background(color)
-                )
-                Spacer(Modifier.width(8.dp))
-                Text(error.tag, fontWeight = FontWeight.Bold)
-            }
-        },
-        text = {
-            Column(
-                modifier = Modifier.verticalScroll(rememberScrollState()),
-                verticalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                Text(
-                    "Time: ${formatTimestamp(error.timestamp)}",
-                    fontSize = 12.sp,
-                    color = MediaColors.TextSecondary
-                )
-                
-                Text(
-                    "Message:",
-                    fontWeight = FontWeight.Bold,
-                    color = MediaColors.TextSecondary,
-                    fontSize = 12.sp
-                )
-                Text(error.message, color = MediaColors.TextPrimary)
-                
-                error.stackTrace?.let { trace ->
-                    Text(
-                        "Stack Trace:",
-                        fontWeight = FontWeight.Bold,
-                        color = MediaColors.TextSecondary,
-                        fontSize = 12.sp
-                    )
-                    Surface(
-                        color = MediaColors.Background,
-                        shape = RoundedCornerShape(4.dp)
-                    ) {
-                        Text(
-                            text = trace,
-                            fontFamily = FontFamily.Monospace,
-                            fontSize = 10.sp,
-                            color = MediaColors.TextPrimary,
-                            modifier = Modifier.padding(8.dp)
-                        )
-                    }
-                }
-            }
-        },
-        confirmButton = {
-            TextButton(onClick = onDismiss) {
-                Text("Close")
-            }
-        },
-        containerColor = MediaColors.BackgroundSurface
-    )
-}
 
 // =============================================================================
 // HELPERS
