+++ CleverFerret/src/main/java/com/universalmedialibrary/debug/ui/ErrorLogItem.kt
@@ -1,0 +1,89 @@
+package com.universalmedialibrary.debug.ui
+
+import android.content.Intent
+import androidx.compose.animation.*
+import androidx.compose.foundation.*
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.hilt.navigation.compose.hiltViewModel
+import com.universalmedialibrary.debug.*
+import com.universalmedialibrary.ui.media.theme.MediaColors
+import java.text.SimpleDateFormat
+import java.util.*
+
+@Composable
+internal fun ErrorLogItem(
+    error: ErrorLog,
+    onClick: () -> Unit
+) {
+    val color = when (error.level) {
+        LogLevel.ERROR -> MediaColors.Error
+        LogLevel.WARNING -> MediaColors.Warning
+        LogLevel.INFO -> MediaColors.AccentPrimary
+    }
+    
+    Surface(
+        modifier = Modifier
+            .fillMaxWidth()
+            .clickable(onClick = onClick),
+        color = color.copy(alpha = 0.1f),
+        shape = RoundedCornerShape(4.dp)
+    ) {
+        Row(
+            modifier = Modifier.padding(8.dp),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Box(
+                modifier = Modifier
+                    .size(8.dp)
+                    .clip(CircleShape)
+                    .background(color)
+            )
+            Spacer(Modifier.width(8.dp))
+            Column(modifier = Modifier.weight(1f)) {
+                Row(
+                    verticalAlignment = Alignment.CenterVertically,
+                    horizontalArrangement = Arrangement.spacedBy(8.dp)
+                ) {
+                    Text(
+                        text = error.tag,
+                        fontSize = 10.sp,
+                        fontWeight = FontWeight.Bold,
+                        fontFamily = FontFamily.Monospace,
+                        color = color
+                    )
+                    Text(
+                        text = formatTimestamp(error.timestamp),
+                        fontSize = 10.sp,
+                        color = MediaColors.TextSecondary
+                    )
+                }
+                Text(
+                    text = error.message,
+                    fontSize = 12.sp,
+                    color = MediaColors.TextPrimary,
+                    maxLines = 2,
+                    overflow = TextOverflow.Ellipsis
+                )
+            }
+        }
+    }
+}
