+++ CleverFerret/src/main/java/com/universalmedialibrary/debug/ui/ErrorLogsTab.kt
@@ -1,0 +1,75 @@
+package com.universalmedialibrary.debug.ui
+
+import android.content.Intent
+import androidx.compose.animation.*
+import androidx.compose.foundation.*
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.hilt.navigation.compose.hiltViewModel
+import com.universalmedialibrary.debug.*
+import com.universalmedialibrary.ui.media.theme.MediaColors
+import java.text.SimpleDateFormat
+import java.util.*
+
+@Composable
+internal fun ErrorLogsTab(
+    errors: List<ErrorLog>,
+    onErrorClick: (ErrorLog) -> Unit,
+    onClearAll: () -> Unit
+) {
+    Column(modifier = Modifier.fillMaxSize()) {
+        if (errors.isNotEmpty()) {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(16.dp),
+                horizontalArrangement = Arrangement.End
+            ) {
+                TextButton(onClick = onClearAll) {
+                    Icon(Icons.Default.Delete, contentDescription = null)
+                    Spacer(Modifier.width(4.dp))
+                    Text("Clear All")
+                }
+            }
+        }
+        
+        if (errors.isEmpty()) {
+            EmptyState(
+                icon = Icons.Default.CheckCircle,
+                message = "No errors logged",
+                color = MediaColors.Success
+            )
+        } else {
+            LazyColumn(
+                modifier = Modifier.fillMaxSize(),
+                contentPadding = PaddingValues(16.dp),
+                verticalArrangement = Arrangement.spacedBy(4.dp)
+            ) {
+                items(errors.reversed(), key = { it.id }) { error ->
+                    ErrorLogItem(
+                        error = error,
+                        onClick = { onErrorClick(error) }
+                    )
+                }
+            }
+        }
+    }
+}
