+++ CleverFerret/src/main/java/com/universalmedialibrary/debug/ui/FlagsTab.kt
@@ -1,0 +1,89 @@
+package com.universalmedialibrary.debug.ui
+
+import android.content.Intent
+import androidx.compose.animation.*
+import androidx.compose.foundation.*
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.hilt.navigation.compose.hiltViewModel
+import com.universalmedialibrary.debug.*
+import com.universalmedialibrary.ui.media.theme.MediaColors
+import java.text.SimpleDateFormat
+import java.util.*
+
+@Composable
+internal fun FlagsTab(
+    flags: Map<String, Boolean>,
+    onToggleFlag: (String, Boolean) -> Unit
+) {
+    LazyColumn(
+        modifier = Modifier.fillMaxSize(),
+        contentPadding = PaddingValues(16.dp),
+        verticalArrangement = Arrangement.spacedBy(8.dp)
+    ) {
+        item {
+            Text(
+                "Feature Flags",
+                fontWeight = FontWeight.Bold,
+                color = MediaColors.TextPrimary,
+                modifier = Modifier.padding(bottom = 8.dp)
+            )
+        }
+        
+        items(flags.entries.toList()) { (key, value) ->
+            Card(
+                colors = CardDefaults.cardColors(containerColor = MediaColors.BackgroundSurface),
+                shape = RoundedCornerShape(8.dp)
+            ) {
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(16.dp),
+                    horizontalArrangement = Arrangement.SpaceBetween,
+                    verticalAlignment = Alignment.CenterVertically
+                ) {
+                    Column(modifier = Modifier.weight(1f)) {
+                        Text(
+                            text = key.replace("_", " ").lowercase()
+                                .replaceFirstChar { it.uppercase() },
+                            fontWeight = FontWeight.Medium,
+                            color = MediaColors.TextPrimary
+                        )
+                        Text(
+                            text = key,
+                            fontSize = 10.sp,
+                            fontFamily = FontFamily.Monospace,
+                            color = MediaColors.TextSecondary
+                        )
+                    }
+                    Switch(
+                        checked = value,
+                        onCheckedChange = { onToggleFlag(key, it) },
+                        colors = SwitchDefaults.colors(
+                            checkedThumbColor = MediaColors.AccentPrimary,
+                            checkedTrackColor = MediaColors.AccentPrimary.copy(alpha = 0.5f)
+                        )
+                    )
+                }
+            }
+        }
+    }
+}
