+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/screens/CategoryCard.kt
@@ -1,0 +1,65 @@
+package com.universalmedialibrary.ui.media.screens
+
+import androidx.compose.animation.*
+import androidx.compose.foundation.*
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.*
+import androidx.compose.foundation.lazy.grid.*
+import androidx.compose.foundation.pager.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material.icons.outlined.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import coil.compose.AsyncImage
+import com.universalmedialibrary.ui.media.components.*
+import com.universalmedialibrary.ui.media.theme.*
+import kotlinx.coroutines.launch
+
+@Composable
+internal fun CategoryCard(
+    category: RadioCategory,
+    onClick: () -> Unit
+) {
+    Surface(
+        modifier = Modifier
+            .width(120.dp)
+            .height(80.dp)
+            .clip(RoundedCornerShape(MediaCorners.MD))
+            .clickable(onClick = onClick),
+        color = category.color.copy(alpha = 0.3f)
+    ) {
+        Box(
+            contentAlignment = Alignment.Center,
+            modifier = Modifier.fillMaxSize()
+        ) {
+            Column(horizontalAlignment = Alignment.CenterHorizontally) {
+                Icon(
+                    imageVector = category.icon,
+                    contentDescription = "${category.name} category",
+                    tint = category.color,
+                    modifier = Modifier.size(28.dp)
+                )
+                Spacer(modifier = Modifier.height(MediaSpacing.XS))
+                Text(
+                    text = category.name,
+                    style = MediaTypography.LabelMedium,
+                    color = MediaColors.TextPrimary,
+                    fontWeight = FontWeight.Medium
+                )
+            }
+        }
+    }
+}
