+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/screens/DownloadedEpisodeItem.kt
@@ -1,0 +1,126 @@
+package com.universalmedialibrary.ui.media.screens
+
+import androidx.compose.animation.*
+import androidx.compose.foundation.*
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.*
+import androidx.compose.foundation.lazy.grid.*
+import androidx.compose.foundation.pager.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material.icons.outlined.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import coil.compose.AsyncImage
+import com.universalmedialibrary.ui.media.components.*
+import com.universalmedialibrary.ui.media.theme.*
+import kotlinx.coroutines.launch
+
+@Composable
+internal fun DownloadedEpisodeItem(
+    episode: PodcastEpisode,
+    onClick: () -> Unit,
+    onPlay: () -> Unit
+) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .clickable(onClick = onClick)
+            .padding(horizontal = MediaSpacing.MD, vertical = MediaSpacing.SM),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        // Artwork
+        Surface(
+            modifier = Modifier
+                .size(56.dp)
+                .clip(RoundedCornerShape(MediaCorners.SM)),
+            color = MediaColors.BackgroundSurface
+        ) {
+            if (episode.showArtworkUrl != null) {
+                AsyncImage(
+                    
+                    model = episode.showArtworkUrl,
+                    contentDescription = "Media image",
+                    contentScale = ContentScale.Crop,
+                    modifier = Modifier.fillMaxSize()
+                )
+            }
+        }
+        
+        Spacer(modifier = Modifier.width(MediaSpacing.MD))
+        
+        Column(modifier = Modifier.weight(1f)) {
+            Text(
+                text = episode.title,
+                style = MediaTypography.BodyMedium,
+                color = MediaColors.TextPrimary,
+                maxLines = 2,
+                overflow = TextOverflow.Ellipsis
+            )
+            
+            Spacer(modifier = Modifier.height(2.dp))
+            
+            Row(
+                verticalAlignment = Alignment.CenterVertically,
+                horizontalArrangement = Arrangement.spacedBy(MediaSpacing.SM)
+            ) {
+                Text(
+                    text = episode.showTitle,
+                    style = MediaTypography.LabelSmall,
+                    color = MediaColors.TextSecondary,
+                    maxLines = 1,
+                    overflow = TextOverflow.Ellipsis
+                )
+                Icon(
+                    imageVector = Icons.Default.DownloadDone,
+                    contentDescription = "Media image",
+                    tint = MediaColors.Success,
+                    modifier = Modifier.size(14.dp)
+                )
+                Text(
+                    text = episode.fileSize ?: "",
+                    style = MediaTypography.LabelSmall,
+                    color = MediaColors.TextTertiary
+                )
+            }
+            
+            if (episode.progress > 0) {
+                Spacer(modifier = Modifier.height(MediaSpacing.XS))
+                LinearProgressIndicator(
+                    progress = { episode.progress },
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .height(2.dp)
+                        .clip(RoundedCornerShape(MediaCorners.Full)),
+                    color = MediaColors.AccentPrimary,
+                    trackColor = MediaColors.ProgressBackground
+                )
+            }
+        }
+        
+        IconButton(onClick = onPlay) {
+            Surface(
+                shape = CircleShape,
+                color = MediaColors.AccentPrimary
+            ) {
+                Icon(
+                    imageVector = Icons.Default.PlayArrow,
+                    contentDescription = "Play",
+                    tint = Color.Black,
+                    modifier = Modifier.padding(MediaSpacing.SM)
+                )
+            }
+        }
+    }
+}
