+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/screens/OldTimeRadioShowCard.kt
@@ -1,0 +1,92 @@
+package com.universalmedialibrary.ui.media.screens
+
+import androidx.compose.animation.*
+import androidx.compose.foundation.*
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.*
+import androidx.compose.foundation.lazy.grid.*
+import androidx.compose.foundation.pager.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material.icons.outlined.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import coil.compose.AsyncImage
+import com.universalmedialibrary.ui.media.components.*
+import com.universalmedialibrary.ui.media.theme.*
+import kotlinx.coroutines.launch
+
+@Composable
+internal fun OldTimeRadioShowCard(
+    show: OldTimeRadioShow,
+    onClick: () -> Unit
+) {
+    Surface(
+        modifier = Modifier
+            .fillMaxWidth()
+            .clip(RoundedCornerShape(MediaCorners.MD))
+            .clickable(onClick = onClick),
+        color = MediaColors.BackgroundElevated
+    ) {
+        Column(
+            modifier = Modifier.padding(MediaSpacing.MD)
+        ) {
+            // Show image
+            Surface(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .aspectRatio(1f)
+                    .clip(RoundedCornerShape(MediaCorners.SM)),
+                color = MediaColors.BackgroundSurface
+            ) {
+                if (show.imageUrl != null) {
+                    AsyncImage(
+                    
+                        model = show.imageUrl,
+                        contentDescription = "${show.title} show image",
+                        contentScale = ContentScale.Crop,
+                        modifier = Modifier.fillMaxSize()
+                    )
+                } else {
+                    Box(contentAlignment = Alignment.Center) {
+                        Icon(
+                            imageVector = Icons.Default.Radio,
+                            contentDescription = "${show.title} radio show",
+                            tint = MediaColors.TextTertiary,
+                            modifier = Modifier.size(48.dp)
+                        )
+                    }
+                }
+            }
+            
+            Spacer(modifier = Modifier.height(MediaSpacing.SM))
+            
+            Text(
+                text = show.title,
+                style = MediaTypography.BodyMedium,
+                color = MediaColors.TextPrimary,
+                maxLines = 2,
+                overflow = TextOverflow.Ellipsis,
+                fontWeight = FontWeight.Medium
+            )
+            
+            Text(
+                text = "${show.episodeCount} episodes",
+                style = MediaTypography.LabelSmall,
+                color = MediaColors.TextSecondary
+            )
+        }
+    }
+}
