+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/screens/PodcastShowCard.kt
@@ -1,0 +1,111 @@
+package com.universalmedialibrary.ui.media.screens
+
+import androidx.compose.animation.*
+import androidx.compose.foundation.*
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.*
+import androidx.compose.foundation.lazy.grid.*
+import androidx.compose.foundation.pager.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material.icons.outlined.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import coil.compose.AsyncImage
+import com.universalmedialibrary.ui.media.components.*
+import com.universalmedialibrary.ui.media.theme.*
+import kotlinx.coroutines.launch
+
+@Composable
+internal fun PodcastShowCard(
+    show: PodcastShow,
+    onClick: () -> Unit
+) {
+    Column(
+        modifier = Modifier
+            .width(150.dp)
+            .clickable(onClick = onClick)
+    ) {
+        Box(
+            modifier = Modifier
+                .size(150.dp)
+                .clip(RoundedCornerShape(MediaCorners.MD))
+        ) {
+            if (show.artworkUrl != null) {
+                AsyncImage(
+                    
+                    model = show.artworkUrl,
+                    contentDescription = show.title,
+                    contentScale = ContentScale.Crop,
+                    modifier = Modifier.fillMaxSize()
+                )
+            } else {
+                Surface(
+                    color = MediaColors.BackgroundSurface,
+                    modifier = Modifier.fillMaxSize()
+                ) {
+                    Icon(
+                        imageVector = Icons.Default.Podcasts,
+                        contentDescription = "Media image",
+                        tint = MediaColors.MediaTypes.Podcast.copy(alpha = 0.5f),
+                        modifier = Modifier
+                            .padding(MediaSpacing.XL)
+                            .fillMaxSize()
+                    )
+                }
+            }
+            
+            // Unplayed badge
+            if (show.unplayedCount > 0) {
+                Box(
+                    modifier = Modifier
+                        .align(Alignment.TopEnd)
+                        .padding(MediaSpacing.XS)
+                ) {
+                    Surface(
+                        shape = CircleShape,
+                        color = MediaColors.AccentPrimary
+                    ) {
+                        Text(
+                            text = if (show.unplayedCount > 99) "99+" else show.unplayedCount.toString(),
+                            style = MediaTypography.LabelSmall,
+                            color = Color.Black,
+                            fontWeight = FontWeight.Bold,
+                            modifier = Modifier.padding(horizontal = MediaSpacing.SM, vertical = 2.dp)
+                        )
+                    }
+                }
+            }
+        }
+        
+        Spacer(modifier = Modifier.height(MediaSpacing.SM))
+        
+        Text(
+            text = show.title,
+            style = MediaTypography.BodyMedium,
+            color = MediaColors.TextPrimary,
+            maxLines = 2,
+            overflow = TextOverflow.Ellipsis,
+            fontWeight = FontWeight.Medium
+        )
+        
+        Text(
+            text = show.author,
+            style = MediaTypography.LabelSmall,
+            color = MediaColors.TextSecondary,
+            maxLines = 1,
+            overflow = TextOverflow.Ellipsis
+        )
+    }
+}
