+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/screens/RadioScreen.kt
@@ -162,141 +162,11 @@ fun RadioScreen(
 // TOP BAR
 // =============================================================================
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun RadioTopBar(
-    onBackClick: () -> Unit,
-    onSearchClick: () -> Unit
-) {
-    TopAppBar(
-        title = {
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                Icon(
-                    imageVector = Icons.Default.Radio,
-                    contentDescription = "Radio",
-                    tint = MediaColors.MediaTypes.Radio,
-                    modifier = Modifier.size(24.dp)
-                )
-                Spacer(modifier = Modifier.width(MediaSpacing.SM))
-                Text(
-                    text = "Radio",
-                    style = MediaTypography.TitleMedium,
-                    color = MediaColors.TextPrimary
-                )
-            }
-        },
-        navigationIcon = {
-            IconButton(onClick = onBackClick) {
-                Icon(
-                    imageVector = Icons.Default.ArrowBack,
-                    contentDescription = "Back",
-                    tint = MediaColors.TextPrimary
-                )
-            }
-        },
-        actions = {
-            IconButton(onClick = onSearchClick) {
-                Icon(
-                    imageVector = Icons.Default.Search,
-                    contentDescription = "Search",
-                    tint = MediaColors.TextSecondary
-                )
-            }
-        },
-        colors = TopAppBarDefaults.topAppBarColors(
-            containerColor = MediaColors.BackgroundElevated
-        )
-    )
-}
 
 // =============================================================================
 // INTERNET RADIO PAGE
 // =============================================================================
 
-@Composable
-private fun InternetRadioPage(
-    state: RadioScreenState,
-    onStationClick: (RadioStation) -> Unit,
-    onFavoriteToggle: (RadioStation) -> Unit,
-    onCategoryClick: (RadioCategory) -> Unit
-) {
-    LazyColumn(
-        contentPadding = PaddingValues(vertical = MediaSpacing.MD),
-        modifier = Modifier.fillMaxSize()
-    ) {
-        // Favorites section
-        if (state.favoriteStations.isNotEmpty()) {
-            item {
-                SectionHeader(title = "Favorites", icon = Icons.Default.Favorite)
-            }
-            item {
-                LazyRow(
-                    contentPadding = PaddingValues(horizontal = MediaSpacing.MD),
-                    horizontalArrangement = Arrangement.spacedBy(MediaSpacing.MD)
-                ) {
-                    items(state.favoriteStations) { station ->
-                        RadioStationCard(
-                            station = station,
-                            onClick = { onStationClick(station) },
-                            onFavorite = { onFavoriteToggle(station) }
-                        )
-                    }
-                }
-            }
-            item { Spacer(modifier = Modifier.height(MediaSpacing.LG)) }
-        }
-        
-        // Categories section
-        item {
-            SectionHeader(title = "Browse by Category", icon = Icons.Default.Category)
-        }
-        item {
-            LazyRow(
-                contentPadding = PaddingValues(horizontal = MediaSpacing.MD),
-                horizontalArrangement = Arrangement.spacedBy(MediaSpacing.MD)
-            ) {
-                items(state.categories) { category ->
-                    CategoryCard(
-                        category = category,
-                        onClick = { onCategoryClick(category) }
-                    )
-                }
-            }
-        }
-        
-        item { Spacer(modifier = Modifier.height(MediaSpacing.LG)) }
-        
-        // Popular stations section
-        item {
-            SectionHeader(title = "Popular Stations", icon = Icons.Default.TrendingUp)
-        }
-        items(state.popularStations) { station ->
-            RadioStationListItem(
-                station = station,
-                isPlaying = state.nowPlaying?.id == station.id,
-                onClick = { onStationClick(station) },
-                onFavorite = { onFavoriteToggle(station) }
-            )
-        }
-        
-        item { Spacer(modifier = Modifier.height(MediaSpacing.LG)) }
-        
-        // Recently played section
-        if (state.recentlyPlayed.isNotEmpty()) {
-            item {
-                SectionHeader(title = "Recently Played", icon = Icons.Default.History)
-            }
-            items(state.recentlyPlayed) { station ->
-                RadioStationListItem(
-                    station = station,
-                    isPlaying = state.nowPlaying?.id == station.id,
-                    onClick = { onStationClick(station) },
-                    onFavorite = { onFavoriteToggle(station) }
-                )
-            }
-        }
-    }
-}
 
 @Composable
 private fun SectionHeader(
@@ -325,252 +195,8 @@ private fun SectionHeader(
     }
 }
 
-@Composable
-private fun RadioStationCard(
-    station: RadioStation,
-    onClick: () -> Unit,
-    onFavorite: () -> Unit
-) {
-    Surface(
-        modifier = Modifier
-            .width(140.dp)
-            .clip(RoundedCornerShape(MediaCorners.MD))
-            .clickable(onClick = onClick),
-        color = MediaColors.BackgroundElevated
-    ) {
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally,
-            modifier = Modifier.padding(MediaSpacing.MD)
-        ) {
-            Box(
-                modifier = Modifier
-                    .size(100.dp)
-                    .clip(CircleShape),
-                contentAlignment = Alignment.Center
-            ) {
-                if (station.logoUrl != null) {
-                    AsyncImage(
-                    
-                        model = station.logoUrl,
-                        contentDescription = "${station.name} logo",
-                        contentScale = ContentScale.Crop,
-                        modifier = Modifier.fillMaxSize()
-                    )
-                } else {
-                    Surface(
-                        color = station.genreColor.copy(alpha = 0.2f),
-                        modifier = Modifier.fillMaxSize()
-                    ) {
-                        Icon(
-                            imageVector = Icons.Default.Radio,
-                            contentDescription = "${station.name} radio station",
-                            tint = station.genreColor,
-                            modifier = Modifier.padding(MediaSpacing.LG)
-                        )
-                    }
-                }
-                
-                // Favorite indicator
-                if (station.isFavorite) {
-                    Box(
-                        modifier = Modifier
-                            .align(Alignment.TopEnd)
-                            .padding(4.dp)
-                    ) {
-                        Icon(
-                            imageVector = Icons.Filled.Favorite,
-                            contentDescription = "Favorite station",
-                            tint = MediaColors.Error,
-                            modifier = Modifier.size(20.dp)
-                        )
-                    }
-                }
-            }
-            
-            Spacer(modifier = Modifier.height(MediaSpacing.SM))
-            
-            Text(
-                text = station.name,
-                style = MediaTypography.BodyMedium,
-                color = MediaColors.TextPrimary,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-                textAlign = TextAlign.Center,
-                fontWeight = FontWeight.Medium
-            )
-            
-            Text(
-                text = station.genre,
-                style = MediaTypography.LabelSmall,
-                color = MediaColors.TextSecondary,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-                textAlign = TextAlign.Center
-            )
-        }
-    }
-}
 
-@Composable
-private fun RadioStationListItem(
-    station: RadioStation,
-    isPlaying: Boolean,
-    onClick: () -> Unit,
-    onFavorite: () -> Unit
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable(onClick = onClick)
-            .background(
-                if (isPlaying) MediaColors.AccentPrimary.copy(alpha = 0.1f) else Color.Transparent
-            )
-            .padding(horizontal = MediaSpacing.MD, vertical = MediaSpacing.SM),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        // Station logo
-        Surface(
-            modifier = Modifier
-                .size(56.dp)
-                .clip(RoundedCornerShape(MediaCorners.SM)),
-            color = station.genreColor.copy(alpha = 0.2f)
-        ) {
-            if (station.logoUrl != null) {
-                AsyncImage(
-                    
-                    model = station.logoUrl,
-                    contentDescription = "${station.name} logo",
-                    contentScale = ContentScale.Crop,
-                    modifier = Modifier.fillMaxSize()
-                )
-            } else {
-                Icon(
-                    imageVector = Icons.Default.Radio,
-                    contentDescription = "${station.name} radio station",
-                    tint = station.genreColor,
-                    modifier = Modifier.padding(MediaSpacing.MD)
-                )
-            }
-            
-            // Playing indicator
-            if (isPlaying) {
-                Box(
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .background(Color.Black.copy(alpha = 0.5f)),
-                    contentAlignment = Alignment.Center
-                ) {
-                    Icon(
-                        imageVector = Icons.Default.VolumeUp,
-                        contentDescription = "Playing",
-                        tint = MediaColors.AccentPrimary,
-                        modifier = Modifier.size(24.dp)
-                    )
-                }
-            }
-        }
-        
-        Spacer(modifier = Modifier.width(MediaSpacing.MD))
-        
-        // Station info
-        Column(modifier = Modifier.weight(1f)) {
-            Text(
-                text = station.name,
-                style = MediaTypography.BodyMedium,
-                color = if (isPlaying) MediaColors.AccentPrimary else MediaColors.TextPrimary,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-                fontWeight = if (isPlaying) FontWeight.SemiBold else FontWeight.Normal
-            )
-            
-            Row(
-                horizontalArrangement = Arrangement.spacedBy(MediaSpacing.SM),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = station.genre,
-                    style = MediaTypography.LabelSmall,
-                    color = MediaColors.TextSecondary
-                )
-                
-                station.country?.let { country ->
-                    Text("•", color = MediaColors.TextTertiary)
-                    Text(
-                        text = country,
-                        style = MediaTypography.LabelSmall,
-                        color = MediaColors.TextTertiary
-                    )
-                }
-                
-                station.bitrate?.let { bitrate ->
-                    Text("•", color = MediaColors.TextTertiary)
-                    Text(
-                        text = "$bitrate kbps",
-                        style = MediaTypography.LabelSmall,
-                        color = MediaColors.TextTertiary
-                    )
-                }
-            }
-            
-            // Current track info
-            station.currentTrack?.let { track ->
-                Spacer(modifier = Modifier.height(2.dp))
-                Text(
-                    text = "♪ $track",
-                    style = MediaTypography.LabelSmall,
-                    color = MediaColors.AccentPrimary,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis
-                )
-            }
-        }
-        
-        // Favorite button
-        IconButton(onClick = onFavorite) {
-            Icon(
-                imageVector = if (station.isFavorite) Icons.Filled.Favorite else Icons.Outlined.FavoriteBorder,
-                contentDescription = if (station.isFavorite) "Remove from favorites" else "Add to favorites",
-                tint = if (station.isFavorite) MediaColors.Error else MediaColors.TextTertiary
-            )
-        }
-    }
-}
 
-@Composable
-private fun CategoryCard(
-    category: RadioCategory,
-    onClick: () -> Unit
-) {
-    Surface(
-        modifier = Modifier
-            .width(120.dp)
-            .height(80.dp)
-            .clip(RoundedCornerShape(MediaCorners.MD))
-            .clickable(onClick = onClick),
-        color = category.color.copy(alpha = 0.3f)
-    ) {
-        Box(
-            contentAlignment = Alignment.Center,
-            modifier = Modifier.fillMaxSize()
-        ) {
-            Column(horizontalAlignment = Alignment.CenterHorizontally) {
-                Icon(
-                    imageVector = category.icon,
-                    contentDescription = "${category.name} category",
-                    tint = category.color,
-                    modifier = Modifier.size(28.dp)
-                )
-                Spacer(modifier = Modifier.height(MediaSpacing.XS))
-                Text(
-                    text = category.name,
-                    style = MediaTypography.LabelMedium,
-                    color = MediaColors.TextPrimary,
-                    fontWeight = FontWeight.Medium
-                )
-            }
-        }
-    }
-}
 
 // =============================================================================
 // FM RADIO PAGE
@@ -674,155 +300,11 @@ private fun OldTimeRadioPage(
     }
 }
 
-@Composable
-private fun OldTimeRadioShowCard(
-    show: OldTimeRadioShow,
-    onClick: () -> Unit
-) {
-    Surface(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clip(RoundedCornerShape(MediaCorners.MD))
-            .clickable(onClick = onClick),
-        color = MediaColors.BackgroundElevated
-    ) {
-        Column(
-            modifier = Modifier.padding(MediaSpacing.MD)
-        ) {
-            // Show image
-            Surface(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .aspectRatio(1f)
-                    .clip(RoundedCornerShape(MediaCorners.SM)),
-                color = MediaColors.BackgroundSurface
-            ) {
-                if (show.imageUrl != null) {
-                    AsyncImage(
-                    
-                        model = show.imageUrl,
-                        contentDescription = "${show.title} show image",
-                        contentScale = ContentScale.Crop,
-                        modifier = Modifier.fillMaxSize()
-                    )
-                } else {
-                    Box(contentAlignment = Alignment.Center) {
-                        Icon(
-                            imageVector = Icons.Default.Radio,
-                            contentDescription = "${show.title} radio show",
-                            tint = MediaColors.TextTertiary,
-                            modifier = Modifier.size(48.dp)
-                        )
-                    }
-                }
-            }
-            
-            Spacer(modifier = Modifier.height(MediaSpacing.SM))
-            
-            Text(
-                text = show.title,
-                style = MediaTypography.BodyMedium,
-                color = MediaColors.TextPrimary,
-                maxLines = 2,
-                overflow = TextOverflow.Ellipsis,
-                fontWeight = FontWeight.Medium
-            )
-            
-            Text(
-                text = "${show.episodeCount} episodes",
-                style = MediaTypography.LabelSmall,
-                color = MediaColors.TextSecondary
-            )
-        }
-    }
-}
 
 // =============================================================================
 // NOW PLAYING BAR
 // =============================================================================
 
-@Composable
-private fun RadioNowPlayingBar(
-    station: RadioStation,
-    onTap: () -> Unit
-) {
-    Surface(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable(onClick = onTap),
-        color = MediaColors.BackgroundElevated,
-        tonalElevation = MediaElevation.MD
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(MediaSpacing.MD),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            // Station logo
-            Surface(
-                modifier = Modifier
-                    .size(48.dp)
-                    .clip(RoundedCornerShape(MediaCorners.SM)),
-                color = station.genreColor.copy(alpha = 0.2f)
-            ) {
-                if (station.logoUrl != null) {
-                    AsyncImage(
-                    
-                        model = station.logoUrl,
-                        contentDescription = "Now playing: ${station.name}",
-                        contentScale = ContentScale.Crop,
-                        modifier = Modifier.fillMaxSize()
-                    )
-                } else {
-                    Icon(
-                        imageVector = Icons.Default.Radio,
-                        contentDescription = "Now playing: ${station.name}",
-                        tint = station.genreColor,
-                        modifier = Modifier.padding(MediaSpacing.SM)
-                    )
-                }
-            }
-            
-            Spacer(modifier = Modifier.width(MediaSpacing.MD))
-            
-            // Station info
-            Column(modifier = Modifier.weight(1f)) {
-                Text(
-                    text = station.name,
-                    style = MediaTypography.BodyMedium,
-                    color = MediaColors.TextPrimary,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                    fontWeight = FontWeight.Medium
-                )
-                
-                station.currentTrack?.let { track ->
-                    Text(
-                        text = track,
-                        style = MediaTypography.LabelSmall,
-                        color = MediaColors.AccentPrimary,
-                        maxLines = 1,
-                        overflow = TextOverflow.Ellipsis
-                    )
-                } ?: Text(
-                    text = station.genre,
-                    style = MediaTypography.LabelSmall,
-                    color = MediaColors.TextSecondary
-                )
-            }
-            
-            // Playback controls
-            IconButton(onClick = { /* Stop */ }) {
-                Icon(
-                    imageVector = Icons.Default.Stop,
-                    contentDescription = "Stop",
-                    tint = MediaColors.TextSecondary
-                )
-            }
-        }
-    }
-}
 
 // =============================================================================
 // EMPTY & UNAVAILABLE STATES
@@ -860,40 +342,6 @@ private fun RadioEmptyState(
     }
 }
 
-@Composable
-private fun RadioUnavailableState(
-    title: String,
-    message: String
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(MediaSpacing.XL),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        Icon(
-            imageVector = Icons.Outlined.SignalCellularNodata,
-            contentDescription = "Radio unavailable",
-            tint = MediaColors.TextTertiary,
-            modifier = Modifier.size(72.dp)
-        )
-        Spacer(modifier = Modifier.height(MediaSpacing.MD))
-        Text(
-            text = title,
-            style = MediaTypography.TitleMedium,
-            color = MediaColors.TextPrimary,
-            textAlign = TextAlign.Center
-        )
-        Spacer(modifier = Modifier.height(MediaSpacing.SM))
-        Text(
-            text = message,
-            style = MediaTypography.BodyMedium,
-            color = MediaColors.TextSecondary,
-            textAlign = TextAlign.Center
-        )
-    }
-}
 
 // =============================================================================
 // DATA MODELS
