+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/screens/RadioUnavailableState.kt
@@ -1,0 +1,64 @@
+package com.universalmedialibrary.ui.media.screens
+
+import androidx.compose.animation.*
+import androidx.compose.foundation.*
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.*
+import androidx.compose.foundation.lazy.grid.*
+import androidx.compose.foundation.pager.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material.icons.outlined.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import coil.compose.AsyncImage
+import com.universalmedialibrary.ui.media.components.*
+import com.universalmedialibrary.ui.media.theme.*
+import kotlinx.coroutines.launch
+
+@Composable
+internal fun RadioUnavailableState(
+    title: String,
+    message: String
+) {
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(MediaSpacing.XL),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center
+    ) {
+        Icon(
+            imageVector = Icons.Outlined.SignalCellularNodata,
+            contentDescription = "Radio unavailable",
+            tint = MediaColors.TextTertiary,
+            modifier = Modifier.size(72.dp)
+        )
+        Spacer(modifier = Modifier.height(MediaSpacing.MD))
+        Text(
+            text = title,
+            style = MediaTypography.TitleMedium,
+            color = MediaColors.TextPrimary,
+            textAlign = TextAlign.Center
+        )
+        Spacer(modifier = Modifier.height(MediaSpacing.SM))
+        Text(
+            text = message,
+            style = MediaTypography.BodyMedium,
+            color = MediaColors.TextSecondary,
+            textAlign = TextAlign.Center
+        )
+    }
+}
