+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/media/screens/RecentSearchesSection.kt
@@ -1,0 +1,94 @@
+package com.universalmedialibrary.ui.media.screens
+
+import androidx.compose.animation.*
+import androidx.compose.foundation.*
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.KeyboardActions
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material.icons.outlined.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.focus.FocusRequester
+import androidx.compose.ui.focus.focusRequester
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalFocusManager
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import coil.compose.AsyncImage
+import com.universalmedialibrary.ui.components.TagChip
+import com.universalmedialibrary.ui.media.components.*
+import com.universalmedialibrary.ui.media.theme.*
+
+@Composable
+internal fun RecentSearchesSection(
+    recentSearches: List<String>,
+    onRecentSearchClick: (String) -> Unit,
+    onClearAll: () -> Unit
+) {
+    Column(modifier = Modifier.fillMaxWidth()) {
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(horizontal = MediaSpacing.MD, vertical = MediaSpacing.SM),
+            horizontalArrangement = Arrangement.SpaceBetween,
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Text(
+                text = "Recent Searches",
+                style = MediaTypography.TitleSmall,
+                color = MediaColors.TextSecondary
+            )
+            
+            TextButton(onClick = onClearAll) {
+                Text(
+                    text = "Clear",
+                    color = MediaColors.AccentSecondary
+                )
+            }
+        }
+        
+        recentSearches.forEach { search ->
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .clickable { onRecentSearchClick(search) }
+                    .padding(horizontal = MediaSpacing.MD, vertical = MediaSpacing.SM),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Icon(
+                    imageVector = Icons.Default.History,
+                    contentDescription = "Media image",
+                    tint = MediaColors.TextTertiary,
+                    modifier = Modifier.size(20.dp)
+                )
+                
+                Spacer(modifier = Modifier.width(MediaSpacing.MD))
+                
+                Text(
+                    text = search,
+                    style = MediaTypography.BodyMedium,
+                    color = MediaColors.TextPrimary,
+                    modifier = Modifier.weight(1f)
+                )
+                
+                Icon(
+                    imageVector = Icons.Default.NorthWest,
+                    contentDescription = "Use this search",
+                    tint = MediaColors.TextTertiary,
+                    modifier = Modifier.size(16.dp)
+                )
+            }
+        }
+    }
+}
