+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/podcast/PodcastManagerScreen.kt
@@ -272,56 +272,6 @@ fun PodcastManagerScreen(
     }
 }
 
-@Composable
-fun PodcastSubscriptionsTab(
-    podcasts: List<Podcast>,
-    onPodcastClick: (Podcast) -> Unit,
-    onUnsubscribe: (Podcast) -> Unit
-) {
-    if (podcasts.isEmpty()) {
-        // Empty state
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(32.dp),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Center
-        ) {
-            Icon(
-                Icons.Default.Podcasts,
-                contentDescription = "Media image",
-                modifier = Modifier.size(80.dp),
-                tint = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.6f)
-            )
-            Spacer(modifier = Modifier.height(16.dp))
-            Text(
-                text = "No Podcast Subscriptions",
-                style = MaterialTheme.typography.headlineSmall,
-                fontWeight = FontWeight.Medium
-            )
-            Text(
-                text = "Search for podcasts or add RSS feeds to get started",
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                modifier = Modifier.padding(top = 8.dp)
-            )
-        }
-    } else {
-        LazyColumn(
-            modifier = Modifier.fillMaxSize(),
-            contentPadding = PaddingValues(16.dp),
-            verticalArrangement = Arrangement.spacedBy(12.dp)
-        ) {
-            items(podcasts) { podcast ->
-                PodcastCard(
-                    podcast = podcast,
-                    onClick = { onPodcastClick(podcast) },
-                    onUnsubscribe = { onUnsubscribe(podcast) }
-                )
-            }
-        }
-    }
-}
 
 @Composable
 fun PodcastEpisodesTab(
@@ -371,409 +321,9 @@ fun PodcastDownloadsTab(
     }
 }
 
-@Composable
-fun PodcastCard(
-    podcast: Podcast,
-    onClick: () -> Unit,
-    onUnsubscribe: () -> Unit
-) {
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable(onClick = onClick),
-        colors = CardDefaults.cardColors(
-            containerColor = MaterialTheme.colorScheme.surface
-        )
-    ) {
-        Row(
-            modifier = Modifier.padding(16.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            // Podcast artwork
-            AsyncImage(
-                    
-                model = podcast.imageUrl ?: null, // Will show podcast icon if no image
-                contentDescription = "Podcast Artwork",
-                modifier = Modifier
-                    .size(80.dp)
-                    .clip(MaterialTheme.shapes.medium)
-                    .background(MaterialTheme.colorScheme.surfaceVariant),
-                contentScale = ContentScale.Crop
-            )
 
-            Spacer(modifier = Modifier.width(16.dp))
 
-            // Podcast info
-            Column(
-                modifier = Modifier.weight(1f)
-            ) {
-                Text(
-                    text = podcast.title,
-                    style = MaterialTheme.typography.titleMedium,
-                    fontWeight = FontWeight.Medium,
-                    maxLines = 2,
-                    overflow = TextOverflow.Ellipsis
-                )
 
-                Spacer(modifier = Modifier.height(4.dp))
-
-                Text(
-                    text = podcast.author ?: "Unknown Author",
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-
-                Spacer(modifier = Modifier.height(4.dp))
-
-                Row(
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Text(
-                        text = "${podcast.totalEpisodes} episodes",
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-
-                    if (podcast.explicit) {
-                        Spacer(modifier = Modifier.width(8.dp))
-                        Surface(
-                            color = MaterialTheme.colorScheme.error,
-                            shape = MaterialTheme.shapes.small
-                        ) {
-                            Text(
-                                text = "E",
-                                style = MaterialTheme.typography.labelSmall,
-                                color = MaterialTheme.colorScheme.onError,
-                                modifier = Modifier.padding(horizontal = 4.dp, vertical = 2.dp)
-                            )
-                        }
-                    }
-                }
-
-                Spacer(modifier = Modifier.height(8.dp))
-
-                Text(
-                    text = podcast.description ?: "No description available",
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    maxLines = 2,
-                    overflow = TextOverflow.Ellipsis
-                )
-            }
-
-            // Actions
-            Column {
-                IconButton(onClick = onUnsubscribe) {
-                    Icon(
-                        Icons.Default.Unsubscribe,
-                        contentDescription = "Unsubscribe",
-                        tint = MaterialTheme.colorScheme.error
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun EpisodeCard(
-    episode: PodcastEpisode,
-    downloadStatus: DownloadStatus?,
-    onClick: () -> Unit,
-    onDownloadClick: () -> Unit,
-    onRetryClick: () -> Unit,
-    onPlayClick: () -> Unit
-) {
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable(onClick = onClick),
-        colors = CardDefaults.cardColors(
-            containerColor = MaterialTheme.colorScheme.surface
-        )
-    ) {
-        Row(
-            modifier = Modifier.padding(12.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            // Play button
-            Surface(
-                modifier = Modifier.size(48.dp),
-                shape = CircleShape,
-                color = MaterialTheme.colorScheme.primary
-            ) {
-                IconButton(onClick = onPlayClick) {
-                    Icon(
-                        Icons.Default.PlayArrow,
-                        contentDescription = "Play",
-                        tint = Color.Black
-                    )
-                }
-            }
-
-            Spacer(modifier = Modifier.width(12.dp))
-
-            // Episode info
-            Column(modifier = Modifier.weight(1f)) {
-                Text(
-                    text = episode.title,
-                    style = MaterialTheme.typography.titleSmall,
-                    fontWeight = FontWeight.Medium,
-                    maxLines = 2,
-                    overflow = TextOverflow.Ellipsis
-                )
-
-                Spacer(modifier = Modifier.height(4.dp))
-
-                Row(
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Text(
-                        text = formatDuration(episode.duration),
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-
-                    Spacer(modifier = Modifier.width(8.dp))
-
-                    Text(
-                        text = SimpleDateFormat("MMM dd", Locale.getDefault()).format(episode.publishDate),
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-
-                    if (episode.isDownloaded) {
-                        Spacer(modifier = Modifier.width(8.dp))
-                        Icon(
-                            Icons.Default.DownloadDone,
-                            contentDescription = "Downloaded",
-                            modifier = Modifier.size(16.dp),
-                            tint = MaterialTheme.colorScheme.primary
-                        )
-                    }
-                }
-
-                Spacer(modifier = Modifier.height(4.dp))
-
-                val readinessText = if (episode.playbackReady) "Playback ready" else "Not ready"
-                val readinessColor = if (episode.playbackReady) {
-                    MaterialTheme.colorScheme.primary
-                } else {
-                    MaterialTheme.colorScheme.error
-                }
-                Text(
-                    text = readinessText,
-                    style = MaterialTheme.typography.labelSmall,
-                    color = readinessColor
-                )
-                episode.playbackFailureReason?.let { reason ->
-                    val warningColor = if (episode.recoveryActionLabel != null) {
-                        MaterialTheme.colorScheme.tertiary
-                    } else {
-                        MaterialTheme.colorScheme.error
-                    }
-                    Text(
-                        text = reason,
-                        style = MaterialTheme.typography.bodySmall,
-                        color = warningColor,
-                        maxLines = 2,
-                        overflow = TextOverflow.Ellipsis
-                    )
-                }
-            }
-
-            when (downloadStatus) {
-                is DownloadStatus.Downloading -> {
-                    CircularProgressIndicator(
-                        progress = { downloadStatus.progress },
-                        modifier = Modifier.size(24.dp),
-                        strokeWidth = 2.dp
-                    )
-                }
-                is DownloadStatus.Queued -> {
-                    Text(
-                        text = "Q${downloadStatus.position}",
-                        style = MaterialTheme.typography.labelSmall
-                    )
-                }
-                is DownloadStatus.Retrying -> {
-                    Text(
-                        text = "Retry ${downloadStatus.attempt}/${downloadStatus.maxRetries}",
-                        style = MaterialTheme.typography.labelSmall,
-                        color = MaterialTheme.colorScheme.error
-                    )
-                }
-                is DownloadStatus.Failed -> {
-                    IconButton(onClick = onRetryClick) {
-                        Icon(
-                            Icons.Default.Refresh,
-                            contentDescription = "Retry download",
-                            tint = MaterialTheme.colorScheme.error
-                        )
-                    }
-                }
-                else -> if (episode.recoveryActionLabel != null) {
-                    Button(onClick = onRetryClick) {
-                        Icon(
-                            Icons.Default.Refresh,
-                            contentDescription = episode.recoveryActionLabel
-                        )
-                        Spacer(modifier = Modifier.width(4.dp))
-                        Text(episode.recoveryActionLabel)
-                    }
-                } else if (!episode.isDownloaded) {
-                    IconButton(onClick = onDownloadClick) {
-                        Icon(
-                            Icons.Default.Download,
-                            contentDescription = "Download"
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun DownloadedEpisodeCard(
-    episode: PodcastEpisode,
-    onClick: () -> Unit,
-    onDeleteClick: () -> Unit
-) {
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable(onClick = onClick),
-        colors = CardDefaults.cardColors(
-            containerColor = MaterialTheme.colorScheme.surface
-        )
-    ) {
-        Row(
-            modifier = Modifier.padding(12.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Icon(
-                Icons.Default.DownloadDone,
-                contentDescription = "Downloaded",
-                modifier = Modifier.size(24.dp),
-                tint = MaterialTheme.colorScheme.primary
-            )
-
-            Spacer(modifier = Modifier.width(12.dp))
-
-            Column(modifier = Modifier.weight(1f)) {
-                Text(
-                    text = episode.title,
-                    style = MaterialTheme.typography.titleSmall,
-                    fontWeight = FontWeight.Medium,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis
-                )
-
-                Text(
-                    text = formatDuration(episode.duration),
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-
-            IconButton(onClick = onDeleteClick) {
-                Icon(
-                    Icons.Default.Delete,
-                    contentDescription = "Delete",
-                    tint = MaterialTheme.colorScheme.error
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun PodcastSearchDialog(
-    searchResults: List<PodcastSearchResult>,
-    isSearching: Boolean,
-    onDismiss: () -> Unit,
-    onSearch: (String) -> Unit,
-    onSubscribe: (PodcastSearchResult) -> Unit
-) {
-    var query by remember { mutableStateOf("") }
-
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { Text("Search Podcasts") },
-        text = {
-            Column {
-                OutlinedTextField(
-                    value = query,
-                    onValueChange = { query = it },
-                    label = { Text("Search") },
-                    placeholder = { Text("Enter podcast name or topic...") },
-                    modifier = Modifier.fillMaxWidth(),
-                    trailingIcon = {
-                        IconButton(onClick = { onSearch(query) }) {
-                            if (isSearching) {
-                                CircularProgressIndicator(modifier = Modifier.size(20.dp))
-                            } else {
-                                Icon(Icons.Default.Search, contentDescription = "Search")
-                            }
-                        }
-                    }
-                )
-
-                Spacer(modifier = Modifier.height(16.dp))
-
-                LazyColumn(
-                    modifier = Modifier.height(300.dp),
-                    verticalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    items(searchResults) { result ->
-                        Card(
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .clickable { onSubscribe(result) }
-                        ) {
-                            Row(
-                                modifier = Modifier.padding(12.dp),
-                                verticalAlignment = Alignment.CenterVertically
-                            ) {
-                                AsyncImage(
-                    
-                                    model = result.imageUrl,
-                                    contentDescription = "Media image",
-                                    modifier = Modifier
-                                        .size(40.dp)
-                                        .clip(MaterialTheme.shapes.small),
-                                    contentScale = ContentScale.Crop
-                                )
-
-                                Spacer(modifier = Modifier.width(12.dp))
-
-                                Column(modifier = Modifier.weight(1f)) {
-                                    Text(
-                                        text = result.title,
-                                        style = MaterialTheme.typography.titleSmall,
-                                        maxLines = 1,
-                                        overflow = TextOverflow.Ellipsis
-                                    )
-                                    Text(
-                                        text = result.author ?: "Unknown Author",
-                                        style = MaterialTheme.typography.bodySmall,
-                                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                                    )
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        },
-        confirmButton = {
-            TextButton(onClick = onDismiss) {
-                Text("Close")
-            }
-        }
-    )
-}
 
 @Composable
 fun AddPodcastFeedDialog(
