+++ CleverFerret/src/main/java/com/universalmedialibrary/ui/podcast/PodcastSubscriptionsTab.kt
@@ -1,0 +1,86 @@
+package com.universalmedialibrary.ui.podcast
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.navigation.NavController
+import coil.compose.AsyncImage
+import com.universalmedialibrary.services.podcast.Podcast
+import com.universalmedialibrary.services.podcast.PodcastEpisode
+import com.universalmedialibrary.services.podcast.PodcastSearchResult
+import com.universalmedialibrary.services.podcast.DownloadStatus
+import com.universalmedialibrary.ui.components.ConfirmationDialog
+import com.universalmedialibrary.ui.components.PinAccessDialog
+import com.universalmedialibrary.ui.theme.MetallicFAB
+import com.universalmedialibrary.ui.theme.MetallicTopAppBar
+import kotlinx.coroutines.flow.collectLatest
+import java.text.SimpleDateFormat
+import java.util.*
+
+@Composable
+fun PodcastSubscriptionsTab(
+    podcasts: List<Podcast>,
+    onPodcastClick: (Podcast) -> Unit,
+    onUnsubscribe: (Podcast) -> Unit
+) {
+    if (podcasts.isEmpty()) {
+        // Empty state
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(32.dp),
+            horizontalAlignment = Alignment.CenterHorizontally,
+            verticalArrangement = Arrangement.Center
+        ) {
+            Icon(
+                Icons.Default.Podcasts,
+                contentDescription = "Media image",
+                modifier = Modifier.size(80.dp),
+                tint = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.6f)
+            )
+            Spacer(modifier = Modifier.height(16.dp))
+            Text(
+                text = "No Podcast Subscriptions",
+                style = MaterialTheme.typography.headlineSmall,
+                fontWeight = FontWeight.Medium
+            )
+            Text(
+                text = "Search for podcasts or add RSS feeds to get started",
+                style = MaterialTheme.typography.bodyMedium,
+                color = MaterialTheme.colorScheme.onSurfaceVariant,
+                modifier = Modifier.padding(top = 8.dp)
+            )
+        }
+    } else {
+        LazyColumn(
+            modifier = Modifier.fillMaxSize(),
+            contentPadding = PaddingValues(16.dp),
+            verticalArrangement = Arrangement.spacedBy(12.dp)
+        ) {
+            items(podcasts) { podcast ->
+                PodcastCard(
+                    podcast = podcast,
+                    onClick = { onPodcastClick(podcast) },
+                    onUnsubscribe = { onUnsubscribe(podcast) }
+                )
+            }
+        }
+    }
+}
